made target icon, fixed window resize issue

+++ SpaceAdventuresGL/src/Novaura/Core/Application.cpp
@@ -88,7 +88,7 @@ namespace Novaura {
 		glViewport(0, 0, width, height);
 		GetWindow()->Width = width;
 		GetWindow()->Height = height;
-		GetWindow()->AspectRatio = width / height;
+		GetWindow()->AspectRatio = (float)width / (float)height;
 
 		m_CameraController->GetCamera().SetProjectionMatrix(width, height);
 	}
