+++ SpaceAdventuresGL/src/Novaura/Core/Window.h
@@ -9,6 +9,6 @@ namespace Novaura {
 		GLFWwindow* Window;
 		float Width = 800.0f;
 		float Height = 600.0f;
-		float AspectRatio = Width / Height;
+		float AspectRatio = (float)Width / (float)Height;
 	};
 }
